requirements.txt
aniso8601==8.1.0
apispec==4.0.0
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
faunadb==4.0.1
Flask==1.1.2
flask-apispec==0.11.0
flask-restplus==0.13.0
flask-restx==0.2.0
future==0.18.2
h2==2.6.2
hpack==3.0.0
hyper==0.7.0
hyperframe==3.2.0
idna==2.10
importlib-metadata==3.4.0
iso8601==0.1.13
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
marshmallow==3.10.0
pyrsistent==0.17.3
python-dotenv==0.15.0
pytz==2020.5
requests==2.25.1
-e git+https://github.com/postrational/rest_api_demo.git@1660892257a41e19f8c7b21c81fbb05be5c188d5#egg=rest_api_demo
-e git+https://github.com/postrational/rest_api_demo.git@1660892257a41e19f8c7b21c81fbb05be5c188d5#egg=shop_faunadb
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.26.2
webargs==7.0.1
Werkzeug==0.16.1
zipp==3.4.0
